pyproject.toml
[project]
name = "mbu_egenbefordring_main_ats"
version = "0.0.1"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "automation-server-client @ git+https://github.com/odense-rpa/automation-server-client.git",
    "mbu-dev-shared-components>=3.0.0",
    "mbu-rpa-core",
    "Pillow == 11.1.0",
    "sqlalchemy >= 2.0.34",
    "openpyxl >= 3.1.5",
    "mbu_msoffice_integration>=1.0.1",
    "pyodbc >= 5.1.0",
    "python-dateutil == 2.9.*",
    "cryptography >= 43.0.0",
    "office365-rest-python-client",
    "requests_ntlm >= 1.2.0",
    "pandas >= 2.2.3",
    "pynput",
    "selenium >= 4.0.0",
]

[tool.uv.sources]
automation-server-client = { git = "https://github.com/odense-rpa/automation-server-client.git", tag = "v0.2.0" }

[tool.setuptools]
py-modules = []
